dasf

+++ libft/ft_atoi.c
@@ -6,13 +6,13 @@
 /*   By: aaudeber <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/05 16:08:21 by aaudeber          #+#    #+#             */
-/*   Updated: 2023/06/20 11:43:19 by aaudeber         ###   ########.fr       */
+/*   Updated: 2023/06/26 13:08:28 by aaudeber         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
 	long	num;
 	int		sign;
